+++ attributes/remi-php70.rb
@@ -8,12 +8,6 @@ when 'fedora'
   # default['yum']['remi-php70']['baseurl'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/php70/$basearch/"
   default['yum']['remi-php70']['mirrorlist'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/php70/$basearch/mirror"
   default['yum']['remi-php70']['description'] = "Remi's PHP 7.0 RPM repository for Fedora Linux #{node['platform_version'].to_i} - $basearch"
-when 'amazon'
-  # Use CentOS 7 repo
-  # default['yum']['remi-php70']['baseurl'] = 'http://cdn.remirepo.net/enterprise/7/php70/$basearch/'
-  default['yum']['remi-php70']['mirrorlist'] = 'http://cdn.remirepo.net/enterprise/7/php70/mirror'
-  default['yum']['remi-php70']['description'] = "Remi's PHP 7.0 RPM repository for Enterprise Linux 7 - $basearch"
-  default['yum']['remi-php70']['priority'] = '9' # amazon repos have priority 10, override for correct php
 when 'rhel'
   # default['yum']['remi-php70']['baseurl'] = "http://cdn.remirepo.net/enterprise/#{node['platform_version'].to_i}/php70/$basearch/"
   default['yum']['remi-php70']['mirrorlist'] = if node['platform_version'].to_i == 7