+++ attributes/remi-php73.rb
@@ -8,12 +8,6 @@ when 'fedora'
   # default['yum']['remi-php73']['baseurl'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/php73/$basearch/"
   default['yum']['remi-php73']['mirrorlist'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/php73/$basearch/mirror"
   default['yum']['remi-php73']['description'] = "Remi's PHP 7.3 RPM repository for Fedora Linux #{node['platform_version'].to_i} - $basearch"
-when 'amazon'
-  # Use CentOS 7 repo
-  # default['yum']['remi-php73']['baseurl'] = 'http://cdn.remirepo.net/enterprise/7/php73/$basearch/'
-  default['yum']['remi-php73']['mirrorlist'] = 'http://cdn.remirepo.net/enterprise/7/php73/mirror'
-  default['yum']['remi-php73']['description'] = "Remi's PHP 7.3 RPM repository for Enterprise Linux 7 - $basearch"
-  default['yum']['remi-php73']['priority'] = '9' # amazon repos have priority 10, override for correct php
 when 'rhel'
   # default['yum']['remi-php73']['baseurl'] = "http://cdn.remirepo.net/enterprise/#{node['platform_version'].to_i}/php73/$basearch/"
   default['yum']['remi-php73']['mirrorlist'] = if node['platform_version'].to_i == 7