+++ attributes/remi-php80.rb
@@ -8,12 +8,6 @@ when 'fedora'
   # default['yum']['remi-php80']['baseurl'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/php80/$basearch/"
   default['yum']['remi-php80']['mirrorlist'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/php80/$basearch/mirror"
   default['yum']['remi-php80']['description'] = "Remi's PHP 8.0 RPM repository for Fedora Linux #{node['platform_version'].to_i} - $basearch"
-when 'amazon'
-  # Use CentOS 7 repo
-  # default['yum']['remi-php80']['baseurl'] = 'http://cdn.remirepo.net/enterprise/7/php80/$basearch/'
-  default['yum']['remi-php80']['mirrorlist'] = 'http://cdn.remirepo.net/enterprise/7/php80/mirror'
-  default['yum']['remi-php80']['description'] = "Remi's PHP 8.0 RPM repository for Enterprise Linux 7 - $basearch"
-  default['yum']['remi-php80']['priority'] = '9' # amazon repos have priority 10, override for correct php
 when 'rhel'
   # default['yum']['remi-php80']['baseurl'] = "http://cdn.remirepo.net/enterprise/#{node['platform_version'].to_i}/php80/$basearch/"
   default['yum']['remi-php80']['mirrorlist'] = if node['platform_version'].to_i == 7