+++ attributes/remi-safe.rb
@@ -8,11 +8,6 @@ when 'fedora'
   # default['yum']['remi-safe']['baseurl'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/remi/$basearch/"
   default['yum']['remi-safe']['mirrorlist'] = "http://cdn.remirepo.net/fedora/#{node['platform_version'].to_i}/safe/$basearch/mirror"
   default['yum']['remi-safe']['description'] = "Safe Remi's RPM repository for Fedora Linux #{node['platform_version'].to_i} - $basearch"
-when 'amazon'
-  # Use CentOS 7 repo
-  # default['yum']['remi-safe']['baseurl'] = 'http://cdn.remirepo.net/enterprise/7/remi/$basearch/'
-  default['yum']['remi-safe']['mirrorlist'] = 'http://cdn.remirepo.net/enterprise/7/safe/mirror'
-  default['yum']['remi-safe']['description'] = "Safe Remi's RPM repository for Enterprise Linux 7 - $basearch"
 when 'rhel'
   # default['yum']['remi-safe']['baseurl'] = "http://cdn.remirepo.net/enterprise/#{node['platform_version'].to_i}/remi/$basearch/"
   default['yum']['remi-safe']['mirrorlist'] = if node['platform_version'].to_i == 7