+++ resources/remi_php73.rb
@@ -20,7 +20,7 @@ action :create do
   yum_remi 'default'
 
   # use repo on C7
-  if rhel_7_or_amazon?
+  if rhel_7?
     yum_repository 'remi-php73' do
       baseurl new_resource.baseurl
       mirrorlist new_resource.mirrorlist
@@ -28,8 +28,6 @@ action :create do
       enabled new_resource.enabled
       gpgcheck new_resource.gpgcheck
       gpgkey new_resource.gpgkey
-      # amazon base repo has priority 10, need to override to get the correct php version
-      priority '9' if amazon?
     end
 
     yum_repository 'remi-php73-debuginfo' do
@@ -38,7 +36,7 @@ action :create do
       enabled new_resource.debug_enabled
       gpgcheck new_resource.gpgcheck
       gpgkey new_resource.gpgkey
-    end
+    end if new_resource.debug_enabled
   else
     # use modules on C8 / Fedora
     yum_remi_modular 'default'