+++ spec/remi_spec.rb
@@ -6,19 +6,19 @@ describe 'yum-remi-chef::remi' do
   default_attributes['yum']['remi-debuginfo']['enabled'] = true
   default_attributes['yum']['remi-debuginfo']['managed'] = true
 
-  context 'on Amazon Linux 2' do
-    platform 'amazon', '2'
+  context 'on CentOS 7' do
+    platform 'centos', '7'
 
     it { is_expected.to create_yum_repository('remi') }
-    it { is_expected.to create_yum_repository('remi-debuginfo') }
+    it { is_expected.to_not create_yum_repository('remi-debuginfo') }
   end
 
-  %w(7 8).each do |version|
-    context "on CentOS #{version}" do
-      platform 'centos', version
+  %w(8 9).each do |version|
+    context "on AlmaLinux #{version}" do
+      platform 'almalinux', version
 
       it { is_expected.to create_yum_repository('remi') }
-      it { is_expected.to create_yum_repository('remi-debuginfo') }
+      it { is_expected.to_not create_yum_repository('remi-debuginfo') }
     end
   end
 
@@ -26,6 +26,6 @@ describe 'yum-remi-chef::remi' do
     platform 'fedora'
 
     it { is_expected.to create_yum_repository('remi') }
-    it { is_expected.to create_yum_repository('remi-debuginfo') }
+    it { is_expected.to_not create_yum_repository('remi-debuginfo') }
   end
 end