+++ test/integration/inspec/controls/remi_modular.rb
@@ -6,11 +6,8 @@ control 'remi-modular' do
     it { should be_enabled }
     its('mirrors') do
       should cmp case os.name
-                 when 'amazon'
-                   'http://cdn.remirepo.net/enterprise/7/modular/mirror'
                  when 'fedora'
                    "http://cdn.remirepo.net/fedora/#{os.release.to_i}/modular/#{os.arch}/mirror"
-
                  else # rhel
                    "http://cdn.remirepo.net/enterprise/#{os.release.to_i}/modular/#{os.arch}/mirror"
                  end
@@ -21,10 +18,8 @@ control 'remi-modular' do
     its('remi-modular.gpgcheck') { should cmp 1 }
     its('remi-modular.gpgkey') do
       should cmp case os.name
-                 when 'amazon'
-                   'https://rpms.remirepo.net/RPM-GPG-KEY-remi'
                  when 'fedora'
-                   'https://rpms.remirepo.net/RPM-GPG-KEY-remi2022'
+                   'https://rpms.remirepo.net/RPM-GPG-KEY-remi2023'
                  else # rhel
                    case os.release.to_i
                    when 8